Task 13/jpa database (#44)
* Added configuration @Slf4j

* Updated InvoiceController.java log.info

* Add TaxCalculator Controller

* delete app/src/main/java/pl/futurecollars/invoicing/controllers/InvoiceApi.java app/src/main/java/pl/futurecollars/invoicing/controllers/InvoiceController.java

* style

* modified checkstyle in Invoice.java

* Modified desc log in TaxCalculatorController.java

* Add tax calculator service and controller

* Add setup SQL with Flyway + changed TestHelper to builder pattern

* Add Tax calculator service tests

* Add Tax calculator service tests

* Refactoring build.gradle and add taxCalculator to TestHelper

* Add h2 database

* Add sql database

* Add sql database

* Add application-sql-pg.yml

* Configured sql database

* Corrected application.yml

* Created jpa database

* Extended Tax Calculator

* Extended Tax Calculator

* Extended Tax Calculator

* Added configuration @Slf4j

* Updated InvoiceController.java log.info

* Add TaxCalculator Controller

* style

* modified checkstyle in Invoice.java

* Add tax calculator service and controller

* Add setup SQL with Flyway + changed TestHelper to builder pattern

* Add Tax calculator service tests

* Add Tax calculator service tests

* Refactoring build.gradle and add taxCalculator to TestHelper

* Add h2 database

* App resources to tests

* Merged differences from task 11

* Prepared test for jpa database

* Added test for jpa database

* Refactoring

* Tests

* Tests

* Merge branch 'task_12/SQL_database' into task_13/JPA_database

---------

+++ app/src/main/java/pl/futurecollars/invoicing/db/Database.java
@@ -10,15 +10,15 @@ import pl.futurecollars.invoicing.model.InvoiceEntry;
 
 public interface Database {
 
-    int save(Invoice invoice);
+    long save(Invoice invoice);
 
-    Optional<Invoice> getByID(int id);
+    Optional<Invoice> getByID(long id);
 
     List<Invoice> getAll();
 
-    Optional<Invoice> update(int id, Invoice updatedInvoice);
+    Optional<Invoice> update(long id, Invoice updatedInvoice);
 
-    Optional<Invoice> delete(int id);
+    Optional<Invoice> delete(long id);
 
     default BigDecimal visit(Predicate<Invoice> predicateInvoice, Function<InvoiceEntry, BigDecimal> invoiceEntryToAmount) {
         return getAll()
